+++ Backend/ProgettoAste2/src/main/java/com/gruppoaste2/progettoaste/dao/PostgresCategoriaDAO.java
@@ -8,7 +8,6 @@ import org.springframework.stereotype.Repository;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -29,19 +28,37 @@ public class PostgresCategoriaDAO implements CategoriaDAO {
                 "VALUES(?, ?)";
         if(jdbcTemplate.update(sql, idCategoria, categoria.getNome()) == 0)
             return null;
+
+        List<AttributoModel> attributi = categoria.getAttributi();
+        if(!attributi.isEmpty())
+            for(AttributoModel attributo : attributi)
+                if(aggiungiAttributoCategoria(idCategoria, attributo) == null)
+                    return null;
+
         return idCategoria;
     }
 
     @Override
-    public int assegnaCategoriaAdOggetto(UUID idOggetto, UUID idCategoria) {
-        final String sql = "INSERT INTO categoria_oggetto(id_oggetto, id_categoria) " +
-                "VALUES(?, ?)";
-        return jdbcTemplate.update(sql, idOggetto, idCategoria);
+    public UUID aggiungiAttributoCategoria(UUID idAttributo, UUID idCategoria, AttributoModel attributo) {
+        final String sql = "INSERT INTO attributo(id, id_categoria, nome) " +
+                "VALUES(?, ?, ?)";
+        if(jdbcTemplate.update(sql, idAttributo, idCategoria, attributo.getNome()) == 0)
+            return null;
+        return idAttributo;
     }
 
     @Override
     public int eliminaCategoria(UUID idCategoria) {
+        if(eliminaAttributiCategoria(idCategoria) == 0)
+            return 0;
+
         final String sql = "DELETE FROM categoria WHERE id = ?";
+        return jdbcTemplate.update(sql, idCategoria);
+    }
+
+    @Override
+    public int eliminaAttributiCategoria(UUID idCategoria) {
+        final String sql = "DELETE FROM attributo WHERE id_categoria = ?";
         return jdbcTemplate.update(sql, idCategoria);
     }
 
@@ -94,10 +111,52 @@ public class PostgresCategoriaDAO implements CategoriaDAO {
         return jdbcTemplate.update(sql, categoriaAggiornata.getNome(), idCategoria);
     }
 
+    @Override
+    public int assegnaCategoriaAdOggetto(UUID idOggetto, UUID idCategoria) {
+        final String sql = "INSERT INTO categoria_oggetto(id_oggetto, id_categoria) " +
+                "VALUES(?, ?)";
+        if(jdbcTemplate.update(sql, idOggetto, idCategoria) == 0)
+            return 0;
+
+        List<AttributoModel> attributi = trovaAttributiCategoria(idCategoria);
+        if(!attributi.isEmpty())
+            for(AttributoModel attributo : attributi)
+                if(assegnaValoreAttributoAdOggetto(idOggetto, attributo) == 0)
+                    return 0;
+
+        return 1;
+    }
+
+    @Override
+    public int rimuoviCategoriaDaOggetto(UUID idOggetto, UUID idCategoria) {
+        List<AttributoModel> attributi = trovaAttributiCategoria(idCategoria);
+        if(!attributi.isEmpty())
+            for(AttributoModel attributo : attributi)
+                if(rimuoviValoreAttributoDaOggetto(idOggetto, attributo.getId()) == 0)
+                    return 0;
+
+        final String sql = "DELETE FROM categoria_oggetto WHERE id_oggetto = ? AND id_categoria = ?";
+        return jdbcTemplate.update(sql, idOggetto, idCategoria);
+    }
+
+    @Override
+    public int assegnaValoreAttributoAdOggetto(UUID idOggetto, AttributoModel attributo) {
+        final String sql = "INSERT INTO attributo_oggetto(id_oggetto, id_attributo, valore) " +
+                "VALUES(?, ?, ?)";
+        return jdbcTemplate.update(sql, idOggetto, attributo.getId(), attributo.getValore());
+    }
+
+    @Override
+    public int rimuoviValoreAttributoDaOggetto(UUID idOggetto, UUID idAttributo) {
+        final String sql = "DELETE FROM attributo_oggetto WHERE id_oggetto = ? AND id_attributo = ?";
+        return jdbcTemplate.update(sql, idOggetto, idAttributo);
+    }
+
     private CategoriaModel makeCategoriaFromResultSet(ResultSet resultSet) throws SQLException {
         UUID idCategoria = UUID.fromString(resultSet.getString("id"));
         String nome = resultSet.getString("nome");
-        return new CategoriaModel(idCategoria, nome, Collections.emptyMap());
+        List<AttributoModel> attributi = trovaAttributiCategoria(idCategoria);
+        return new CategoriaModel(idCategoria, nome, attributi);
     }
 
     private AttributoModel makeAttributoFromResultSet(ResultSet resultSet) throws SQLException {
